feat: docker

+++ Mercatura/src/main/java/com/felipemcassiano/Mercatura/models/product/ProductPriceRangeFilterDTO.java
@@ -2,5 +2,5 @@ package com.felipemcassiano.Mercatura.models.product;
 
 import jakarta.annotation.Nullable;
 
-public record ProductPriceRangeFilterDTO(@Nullable Integer min, @Nullable Integer max) {
+public record ProductPriceRangeFilterDTO(@Nullable Long min, @Nullable Long max) {
 }
